readme: update individual ext readmes
Fixes #6518

The following commit updates the readme's of each individual ext
present in the main repository to follow the new template proposed from
the `getting-started` update.

The readmes are updated to use a uniform template, a brief description
of their functionality, useful links (ex: generated API docs), and
the Eclipse license and Theia trademark.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>